refactor: View, Rest 컨트롤러 분리
